pyproject.toml
[project]
name = "gemini-media-mcp"
dynamic = ["version"]
description = "MCP server for Gemini media generation"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.23.1",
    "google-genai>=1.53.0",
    "pillow>=12.0.0",
    "aiohttp>=3.13.2",
    "google-cloud-storage>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/CxOAGI/gemini-media-mcp"
Repository = "https://github.com/CxOAGI/gemini-media-mcp"
Issues = "https://github.com/CxOAGI/gemini-media-mcp/issues"

[project.scripts]
gemini-media-mcp = "src.__main__:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-timeout>=2.4.0",
    "licensecheck>=2025.1.0",
    "basedpyright>=1.25.0",
    "ruff>=0.14.8",
    "prek>=0.2.19",
]
